requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.1
requests==2.31.0
